relacion entity
