--- back/src/modules/payment-order/payment-order.entity.ts
@@ -1,13 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column } from 'typeorm';
-
-@Entity({ name: 'paymentOrders' })
-export class PaymentOrder {
-  @PrimaryGeneratedColumn('uuid')
-  id: string;
-
-  @Column()
-  date: Date;
-}
-
-// una vez que hablemos con el profesor vamos a crear las relaciones
-//estudiantes id
